Refactor code transforms and vault operations out of main
